ai-service/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ===== LangGraph AI Assistant Dependencies =====
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
httpx>=0.27.0
sse-starlette>=1.6.0
